Move Reviewable into Commentable
The Review domain object inherits from the Comment domain object.

Because of that I want the Reviewable concern to be closely linked to the Commentable concern.

There's no situation I can foresee that would include adding a Review without adding a comment, because they are "locked together" at the database level.